pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "jps-yaml-schema-validator"
version = "0.0.0"
description = "A flexible validator that uses YAML rule files to validate user configuration files with type constraints and semantic checks."
readme = "README.md"
authors = [{ name = "Jaideep Sundaram", email = "jai.python3@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["yaml-schema-validator", "bootstrap", "project-generator", "automation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
    "PyYAML==6.0.2",
    "rich==13.9.2",
    "ruamel.yaml==0.18.6",
    "typer==0.12.4",
    "click==8.1.7",
]

[project.urls]
Homepage = "https://github.com/jai-python3/jps-yaml-schema-validator"
Repository = "https://github.com/jai-python3/jps-yaml-schema-validator"
Issues = "https://github.com/jai-python3/jps-yaml-schema-validator/issues"

[project.scripts]
jps-yaml-schema-validate = "jps_yaml_schema_validator.cli:main"

[project.optional-dependencies]
dev = [
    "flake8>=7.0.0",
    "black>=24.0.0",
    "build>=1.2.1",
    "twine>=5.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "isort>=5.13.0",
    "codecov>=2.1.13",
    "autoflake>=2.3.1",
    "pre-commit>=3.8.0",
    "bandit>=1.7.9",
    "vulture>=2.11",
    "flynt>=1.0.1",
    "pydocstyle>=6.3.0",
    "darglint>=1.8.1",
    "mypy>=1.12.1"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

# -----------------------------
# Formatting Tools
# -----------------------------
[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

# -----------------------------
# Testing & Coverage
# -----------------------------
[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-v --disable-warnings"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true

# -----------------------------
# Type Checking
# -----------------------------
[tool.mypy]
python_version = "3.12"

# --- Strictness for production code ---
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_unreachable = true
warn_return_any = true
show_error_codes = true

# --- Paths to exclude from main analysis ---
exclude = [
    "build/",
    "dist/",
]

# --- Override: Relax typing rules for test modules ---
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false
ignore_missing_imports = true

# -----------------------------
# Linting
# -----------------------------
[tool.flake8]
max-line-length = 120

# E203: Black disagrees with slice whitespace rules
# W503: Black prefers line breaks before binary operators
# W292: Black intentionally removes trailing newline at EOF (Black behavior)
extend-ignore = ["E203", "W503", "W292"]

exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
    ".venv",
    ".tox",
    ".pytest_cache",
    "docs"
]

# -----------------------------
# Security (Bandit)
# -----------------------------
[tool.bandit]
# Rules intentionally skipped because they are either safe in this context or false positives.
# Skipped rules:
#   B101 - assert_used: Safe since assertions are not used for security enforcement.
#   B110 - try_except_pass: Handled via code review standards instead.
#   B404 - import_subprocess: subprocess used safely for CLI operations (no user input).
#   B603 - subprocess_without_shell_equals_true: All subprocess calls use shell=False.
#   B607 - start_process_with_partial_path: Safe usage of 'gh' (GitHub CLI) in developer tools.
#   B108 - hardcoded_tmp_directory: Acceptable since temp dirs are user- and timestamp-specific.

skips = ["B101", "B110", "B404", "B603", "B607", "B108"]

# Bandit target directories (recursive)
targets = ["src"]
recursive = true

# Exclude test code from Bandit scans
exclude = ["tests"]

# -----------------------------
# Docstring Style Checkers
# -----------------------------
[tool.pydocstyle]
convention = "google"
add-ignore = [
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring
]
match-dir = "src"

[tool.darglint]
docstring_style = "google"
strictness = "short"
ignore_regex = "(^test_)"
message_template = "{path}:{line}: {msg_id} {msg}"

